+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.2.2
+- Add `--backfill` option to model migration
+- Handle legacy data (that doesn't have log data)
+
 # 0.2.1
 - Support both Rails 4 and 5